adding additional compiler warnings

+++ src/parsers/png.h
@@ -3,4 +3,4 @@
 #include <stddef.h>
 #include "../texture.h"
 
-texture_t* parse_png(const unsigned char* buffer, size_t size);
+texture_t* parse_png(const unsigned char* buffer, uint32_t size);
